requirements.txt
paho-mqtt>=1.6
pyyaml>=6.0
